Refactoring Skills for single-step control

+++ Assets/LostGeneration/Scripts/Model/Board.cs
@@ -9,7 +9,7 @@ namespace LostGen {
         public bool IsSolid;
         public bool IsOpaque;
         public bool IsDiggable;
-        public int BlockType;
+        public byte BlockType;
     }
 
     public class Board {
